alterações

+++ PblSolution/Pbl/Models/ViewModel/ListagemMedsAlunoViewModel.cs
@@ -10,7 +10,7 @@ namespace Pbl.Models.ViewModel
         public int idMed { get; set; }
         public string descMed { get; set; }
         public string descSemestre { get; set; }
-        public int[] notas { get; set; }
+        public double[] notas { get; set; }
         public int[] idControleNotas { get; set; }
     }
 }
